# openapi.yml
openapi: 3.0.3
info:
  title: Resourcery API
  description: API documentation for the Resourcery community resource booking system.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Spring Boot — Auth Service
  - url: http://localhost:8081
    description: PHP — Resources Service
  - url: http://localhost:8082
    description: Django — Bookings & Notifications Service
...
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login successful, JWT cookie set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid credentials
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Self-register a new resident account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Account created, JWT cookie set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Username taken or invalid
  /api/auth/admin/create-member:
    post:
      tags: [Auth]
      summary: Admin creates a member account (no session change)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Member created
        '400':
          description: Username taken or invalid
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current session user
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
  /api/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, currentPassword, newPassword]
              properties:
                username: { type: string }
                currentPassword: { type: string }
                newPassword: { type: string }
      responses:
        '200':
          description: Password updated
        '401':
          description: Wrong current password
...
  /api/auth/members/{id}:
    delete:
      tags: [Auth]
      summary: Delete a member (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Member deleted
        '404':
          description: Not found

# ── RESOURCE ENDPOINTS (PHP) ──────────────────────────────────────────────────
...
components:
  schemas:
    UserResponse:
      type: object
      properties:
        username: { type: string }
        role: { type: string, enum: [ROLE_USER, ROLE_ADMIN] }
        email: { type: string }
...
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        role: { type: string, enum: [ROLE_USER, ROLE_ADMIN], default: ROLE_USER }
...